requirements.txt
estroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.4
decorator==4.3.2
Flask==1.0.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
more-itertools==5.0.0
networkx==2.2
numpy==1.16.0
Pillow==5.4.1
pluggy==0.8.1
py==1.7.0
pydicom==1.2.2
pylint==2.2.2
pyparsing==2.3.1
pyserial==3.4
pytest==4.1.1
python-dateutil==2.7.5
PyWavelets==1.0.2
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
toolz==0.9.0
Werkzeug==0.14.1
wrapt==1.11.1
